+++ components/auth/AuthProvider.tsx
@@ -1,28 +1,39 @@
 "use client"
 
-import type React from "react"
-
-import { useEffect, useState } from "react"
+import { useEffect, useState } from "react" // Changed from "import * as React from "react""
+// import { useRouter } from "next/navigation" // Removed useRouter
+import { useTranslations } from "next-intl" // Added for t()
 import { supabase } from "@/lib/supabase-client"
 import { useAppStore } from "@/lib/store"
 import type { User } from "@/types"
-import { AuthModal } from "./AuthModal"
-import { ApiKeySetupGuide } from "./ApiKeySetupGuide"
+import { AuthModal } from "./AuthModal" // Ensured PascalCase filename
+import { ApiKeySetupGuide } from "./ApiKeySetupGuide" // Ensured PascalCase filename
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
-  const { user, setUser, setLoading } = useAppStore()
-  const [showAuth, setShowAuth] = useState(false)
-  const [showApiKeySetup, setShowApiKeySetup] = useState(false)
+  // const router = useRouter() // Removed useRouter
+  const t = useTranslations() // Added for t()
+  const { user, setUser, fetchInitialData, settings } = useAppStore()
+  const [showAuth, setShowAuth] = useState(false) // Changed from React.useState
+  const [showApiKeySetup, setShowApiKeySetup] = useState(false) // Changed from React.useState
+  const [isLoading, setIsLoading] = useState(true) // Changed from React.useState
 
-  useEffect(() => {
+  useEffect(() => { // Changed from React.useEffect
     // Get initial session
     supabase.auth.getSession().then(({ data: { session } }) => {
       if (session?.user) {
-        fetchUserProfile(session.user.id)
+        setUser(session.user as User)
       } else {
-        setShowAuth(true)
+        // Sign in anonymously for guest users
+        supabase.auth.signInAnonymously().then(({ data, error }) => {
+          if (error) {
+            console.error("Error signing in anonymously:", error)
+            setShowAuth(true)
+          } else if (data.user) {
+            setUser(data.user as User)
+          }
+        })
       }
-      setLoading(false)
+      setIsLoading(false)
     })
 
     // Listen for auth changes
@@ -30,7 +41,27 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
       data: { subscription },
     } = supabase.auth.onAuthStateChange(async (event, session) => {
       if (event === "SIGNED_IN" && session?.user) {
-        await fetchUserProfile(session.user.id)
+        const newUser = session.user as User
+        setUser(newUser)
+
+        // Check for guest data migration
+        const guestUserId = sessionStorage.getItem("guestUserId")
+        if (guestUserId && guestUserId !== newUser.id) {
+          try {
+            await fetch("/api/migrate-guest-data", {
+              method: "POST",
+              headers: { "Content-Type": "application/json" },
+              body: JSON.stringify({
+                new_user_id: newUser.id,
+                guest_user_id: guestUserId,
+              }),
+            })
+            sessionStorage.removeItem("guestUserId")
+          } catch (error) {
+            console.error("Error migrating guest data:", error)
+          }
+        }
+
         setShowAuth(false)
       } else if (event === "SIGNED_OUT") {
         setUser(null)
@@ -39,32 +70,59 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
     })
 
     return () => subscription.unsubscribe()
-  }, [])
+  }, [setUser]) // Changed from React.useEffect
 
-  useEffect(() => {
-    if (user && !user.gemini_api_key) {
+  useEffect(() => { // Changed from React.useEffect
+    if (user) {
+      fetchInitialData()
+    }
+  }, [user, fetchInitialData])
+
+  useEffect(() => { // Changed from React.useEffect
+    // Show API key setup if user is authenticated (not anonymous) and has no API key
+    if (user && !user.is_anonymous && settings && !settings.gemini_api_key) {
       setShowApiKeySetup(true)
+    } else {
+      setShowApiKeySetup(false)
     }
-  }, [user])
+  }, [user, settings])
 
-  const fetchUserProfile = async (userId: string) => {
-    try {
-      const { data, error } = await supabase.from("users").select("*").eq("id", userId).single()
-
-      if (error) throw error
-      setUser(data as User)
-    } catch (error) {
-      console.error("Error fetching user profile:", error)
+  const handleGuestToAuth = () => {
+    if (user?.is_anonymous) {
+      sessionStorage.setItem("guestUserId", user.id)
     }
+    setShowAuth(true)
+  }
+
+  if (isLoading) {
+    return (
+      <div className="min-h-screen flex items-center justify-center">
+        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-primary"></div>
+      </div>
+    )
   }
 
   if (showAuth) {
-    return <AuthModal />
+    return <AuthModal onGuestContinue={() => setShowAuth(false)} onClose={() => setShowAuth(false)} />
   }
 
   if (showApiKeySetup) {
-    return <ApiKeySetupGuide onComplete={() => setShowApiKeySetup(false)} />
+    return <ApiKeySetupGuide onComplete={() => setShowApiKeySetup(false)} onSkip={() => setShowApiKeySetup(false)} />
   }
 
-  return <>{children}</>
+  return (
+    <>
+      {children}
+      {user?.is_anonymous && (
+        <div className="fixed bottom-4 right-4 z-50">
+          <button
+            onClick={handleGuestToAuth}
+            className="bg-primary text-primary-foreground px-4 py-2 rounded-lg shadow-lg hover:bg-primary/90 transition-colors"
+          >
+            {t("auth.signInWithAccount")} {/* Changed from hardcoded text */}
+          </button>
+        </div>
+      )}
+    </>
+  )
 }
